requirements.txt
aerich==0.6.3
aiomysql==0.1.1
aioredis==2.0.1
aiosmtplib==1.1.6
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
Deprecated==1.2.13
dictdiffer==0.9.0
dnspython==2.2.1
email-validator==1.1.3
fakeredis==1.7.1
fastapi==0.75.0
fastapi-mail==1.0.6
h11==0.12.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.22.0
idna==3.3
iso8601==0.1.16
Jinja2==3.0.3
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
Pillow==9.1.1
pycparser==2.21
pydantic==1.9.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.7
pypika-tortoise==0.1.3
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
redis==4.1.4
rfc3986==1.5.0
sh==1.14.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sse-starlette==0.10.3
starlette==0.17.1
tomlkit==0.11.0
tortoise-orm==0.18.1
typing_extensions==4.1.1
uvicorn==0.17.6
watchgod==0.8
websockets==10.2
wrapt==1.14.0
Flask==1.1.2
h11==0.12.0
imutils==0.5.4
itsdangerous==1.1.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-vlc==3.0.11115
qimage2ndarray==1.8.3
Werkzeug==1.0.1